+++ src/game_logic.c
@@ -383,8 +383,6 @@ void game_over(GameState *state)
     // Wait a little without taking input so the user does not accidentally exit at the end
     wait_usec(750000);  // 0.75 seconds
     flush_stdin();
-    #ifdef _WIN32
-    while(input_available()) getchar(); // On Windows, stdin does not necessarily is fully cleared right after the previous function returns
-    #endif // _Win32
-    getchar();  // Wait for the user to press any key to exit
+    int input = getchar();  // Wait for the user to press any key to exit
+    flush_stdin();  // So any remaining input isn't echoed on the terminal
 }